Fixed problem where register_to_payment caused the loading framework to die. Also fixed all calls to the removed django.models.loading.
