+++ .travis.yml
@@ -6,6 +6,8 @@ env:
   - DJANGO_VERSION=1.6
   - DJANGO_VERSION=1.7
   - DJANGO_VERSION=1.8.3
+  - DJANGO_VERSION=1.8.5
+  - DJANGO_VERSION=1.9b1
 # command to install dependencies
 install: 
   - "pip install -q Django==$DJANGO_VERSION"
