+++ getpaid/__init__.py
@@ -1,4 +1,2 @@
-#noinspection PyUnresolvedReferences
-from .models import register_to_payment
-
+from .utils import register_to_payment
 default_app_config = 'getpaid.apps.Config'
