+++ getpaid/apps.py
@@ -1,6 +1,5 @@
 from django.apps import AppConfig
 
-
 class Config(AppConfig):
     name = 'getpaid'
     verbose_name = 'application getpaid'
