+++ getpaid/backends/epaydk/__init__.py
@@ -13,7 +13,7 @@ from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
 from django.utils.translation import get_language_from_request
 from django.contrib.sites.models import Site
-from django.db.models.loading import get_model
+from django.apps import apps
 from django.core.exceptions import ImproperlyConfigured
 from django.utils.six.moves.urllib.parse import urlparse, urlunparse,\
     parse_qsl, urlencode, quote
@@ -238,7 +238,7 @@ class PaymentProcessor(PaymentProcessorBase):
         """
         Payment was confirmed.
         """
-        Payment = get_model('getpaid', 'Payment')
+        Payment = apps.get_model('getpaid', 'Payment')
         with commit_on_success_or_atomic():
             payment = Payment.objects.get(id=params['orderid'])
             assert payment.status == 'accepted_for_proc',\
@@ -256,7 +256,7 @@ class PaymentProcessor(PaymentProcessorBase):
         """
         Payment was accepted into the queue for processing.
         """
-        Payment = get_model('getpaid', 'Payment')
+        Payment = apps.get_model('getpaid', 'Payment')
         with commit_on_success_or_atomic():
             payment = Payment.objects.get(id=payment_id)
             assert payment.status == 'in_progress',\
@@ -268,7 +268,7 @@ class PaymentProcessor(PaymentProcessorBase):
         """
         Payment was cancelled.
         """
-        Payment = get_model('getpaid', 'Payment')
+        Payment = apps.get_model('getpaid', 'Payment')
         with commit_on_success_or_atomic():
             payment = Payment.objects.get(id=payment_id)
             payment.change_status('cancelled')
