+++ getpaid/backends/moip/__init__.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 import logging
 import datetime
 from django.core.urlresolvers import reverse
-from django.db.models import get_model
+from django.apps import apps
 from django.utils.timezone import utc
 import requests
 import time
@@ -94,7 +94,7 @@ class PaymentProcessor(PaymentProcessorBase):
 
     @staticmethod
     def process_notification(params):
-        Payment = get_model('getpaid', 'Payment')
+        Payment = apps.get_model('getpaid', 'Payment')
         try:
             payment = Payment.objects.get(pk=int(params["id"].split("-")[0]))
         except Payment.DoesNotExist:
