+++ getpaid/backends/payu/tasks.py
@@ -1,6 +1,6 @@
 import logging
 from celery.task.base import task
-from django.db.models.loading import get_model
+from django.apps import apps
 
 
 logger = logging.getLogger('getpaid.backends.payu')
@@ -8,7 +8,7 @@ logger = logging.getLogger('getpaid.backends.payu')
 
 @task(max_retries=50, default_retry_delay=2*60)
 def get_payment_status_task(payment_id, session_id):
-    Payment = get_model('getpaid', 'Payment')
+    Payment = apps.get_model('getpaid', 'Payment')
     try:
         payment = Payment.objects.get(pk=int(payment_id))
     except Payment.DoesNotExist:
@@ -21,7 +21,7 @@ def get_payment_status_task(payment_id, session_id):
 
 @task(max_retries=50, default_retry_delay=2 * 60)
 def accept_payment(payment_id, session_id):
-    Payment = get_model('getpaid', 'Payment')
+    Payment = apps.get_model('getpaid', 'Payment')
     try:
         payment = Payment.objects.get(pk=int(payment_id))
     except Payment.DoesNotExist:
