+++ getpaid/backends/przelewy24/tasks.py
@@ -1,13 +1,13 @@
 import logging
 from celery.task.base import task
-from django.db.models.loading import get_model
+from django.apps import apps
 
 logger = logging.getLogger('getpaid.backends.przelewy24')
 
 
 @task
 def get_payment_status_task(payment_id, p24_session_id, p24_order_id, p24_kwota):
-    Payment = get_model('getpaid', 'Payment')
+    Payment = apps.get_model('getpaid', 'Payment')
     try:
         payment = Payment.objects.get(pk=int(payment_id))
     except Payment.DoesNotExist:
