+++ getpaid_test_project/getpaid_test_project/orders/templates/orders/order_detail.html
@@ -1,4 +1,3 @@
-{% load url from future %}
 <html>
 <body>
 <h1>Order #{{ object.pk }}</h1>
